pyproject.toml
[tool.poetry]
name = "kindleanki"
version = "0.0.1"
description = "Kindle to Anki exporter"
authors = ["Michał Kowol <michal.kowol@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.9"
requests = "2.28.2"
deepl = "1.14.0"
attrs = "22.2.0"
dataclasses-json = "0.5.7"
alive-progress = "3.0.1"

[tool.poetry.scripts]
app = "app:main"
test = "scripts:test"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
